refactor: refactor reusable constants
